fixed doubled getUsers call (again)
