+++ client/pages/AssessmentHome/FraHome/Collaborators/Collaborators.tsx
@@ -3,7 +3,6 @@ import React, { useEffect } from 'react'
 import { Users } from '@meta/user'
 
 import { useAppDispatch } from '@client/store'
-import { useAssessment, useCycle } from '@client/store/assessment'
 import { useUser } from '@client/store/user'
 import { UserManagementActions } from '@client/store/userManagement'
 import { useUsers, useUserToEdit } from '@client/store/userManagement/hooks'
@@ -16,8 +15,6 @@ const Collaborators: React.FC = () => {
   const dispatch = useAppDispatch()
 
   const countryIso = useCountryIso()
-  const assessment = useAssessment()
-  const cycle = useCycle()
   const user = useUser()
 
   const userToEdit = useUserToEdit()
@@ -26,13 +23,6 @@ const Collaborators: React.FC = () => {
   useEffect(() => {
     dispatch(UserManagementActions.setUserToEdit(null))
   }, [dispatch])
-
-  useEffect(() => {
-    if (user)
-      dispatch(
-        UserManagementActions.getUsers({ countryIso, assessmentName: assessment.props.name, cycleName: cycle.name })
-      )
-  }, [countryIso, cycle, assessment, user, dispatch])
 
   return userToEdit ? (
     <EditUserForm user={userToEdit} />
