+++ client/pages/AssessmentHome/FraHome/FraHome.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Navigate, NavLink, Route, Routes } from 'react-router-dom'
 
@@ -6,6 +6,10 @@ import classNames from 'classnames'
 
 import { Areas } from '@meta/area'
 
+import { useAppDispatch } from '@client/store'
+import { useAssessment, useCycle } from '@client/store/assessment'
+import { useUser } from '@client/store/user'
+import { UserManagementActions } from '@client/store/userManagement'
 import { useCountryIso } from '@client/hooks'
 import { AssessmentHomeRouteNames } from '@client/basePaths'
 
@@ -16,11 +20,23 @@ import SelectedCountries from './SelectedCountries'
 
 const FraHome: React.FC = () => {
   const { i18n } = useTranslation()
+  const dispatch = useAppDispatch()
   const countryIso = useCountryIso()
+  const assessment = useAssessment()
+  const cycle = useCycle()
+  const user = useUser()
   const sections = useSections()
 
   // tabs are available when user is logged-in and selected area is country
   const displayTabs = sections.length > 1 && Areas.isISOCountry(countryIso)
+
+  useEffect(() => {
+    if (user) {
+      dispatch(
+        UserManagementActions.getUsers({ countryIso, assessmentName: assessment.props.name, cycleName: cycle.name })
+      )
+    }
+  }, [countryIso, cycle, assessment, user, dispatch])
 
   return (
     <>
